+++ src/test/java/org/sbrubbles/conditio/RestartTest.java
@@ -2,7 +2,14 @@ package org.sbrubbles.conditio;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.sbrubbles.conditio.fixtures.BasicCondition;
+import org.sbrubbles.conditio.policies.Policies;
+import org.sbrubbles.conditio.policies.ReturnTypePolicy;
 import org.sbrubbles.conditio.restarts.UseValue;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -19,11 +26,11 @@ public class RestartTest {
   @Test
   public void nullParametersAreNotAllowed() {
     assertThrows(NullPointerException.class,
-      () -> Restarts.on(null, this::body), "missing optionType");
+        () -> Restarts.on(null, this::body), "missing optionType");
     assertThrows(NullPointerException.class,
-      () -> Restarts.on(A.class, null), "missing body");
+        () -> Restarts.on(A.class, null), "missing body");
     assertThrows(NullPointerException.class,
-      () -> Restarts.on(null, null), "missing both");
+        () -> Restarts.on(null, null), "missing both");
   }
 
   @Test
@@ -47,39 +54,80 @@ public class RestartTest {
   @Test
   public void apply() {
     assertEquals(
-      "OK: OMGWTFBBQ",
-      rA.apply(new A("OMGWTFBBQ")));
+        "OK: OMGWTFBBQ",
+        rA.apply(new A("OMGWTFBBQ")));
 
     assertEquals(
-      "OK: OMGWTFBBQ",
-      rA.apply(new B("OMGWTFBBQ")));
+        "OK: OMGWTFBBQ",
+        rA.apply(new B("OMGWTFBBQ")));
 
     assertEquals(
-      "FAIL!",
-      rA.apply(new A("FAIL")));
+        "FAIL!",
+        rA.apply(new A("FAIL")));
 
     assertEquals(
-      "FAIL!",
-      rA.apply(new B("FAIL")));
+        "FAIL!",
+        rA.apply(new B("FAIL")));
   }
 
   @Test
   public void applyForB() {
     assertThrows(
-      ClassCastException.class,
-      () -> rB.apply(new A("OMGWTFBBQ")));
+        ClassCastException.class,
+        () -> rB.apply(new A("OMGWTFBBQ")));
 
     assertThrows(
-      ClassCastException.class,
-      () -> rB.apply(new UseValue<>("OMGWTFBBQ")));
+        ClassCastException.class,
+        () -> rB.apply(new UseValue<>("OMGWTFBBQ")));
 
     assertEquals(
-      "OK: OMGWTFBBQ",
-      rB.apply(new B("OMGWTFBBQ")));
+        "OK: OMGWTFBBQ",
+        rB.apply(new B("OMGWTFBBQ")));
 
     assertEquals(
-      "FAIL!",
-      rB.apply(new B("FAIL")));
+        "FAIL!",
+        rB.apply(new B("FAIL")));
+  }
+
+  @Test
+  public void restartNotFoundExceptionIsThrownByDefault() {
+    try (Scope a = Scopes.create()) {
+      a.handle(BasicCondition.class, (s, ops) -> ops.restart(new UseValue<>(1)));
+
+      a.signal(new BasicCondition(""), new Policies<>(ReturnTypePolicy.ignore()));
+    } catch(RestartNotFoundException r) {
+      assertEquals(new UseValue<>(1), r.getOption());
+
+      return;
+    }
+
+    fail("Should've thrown a RestartNotFoundException earlier!");
+  }
+
+  @Test
+  public void restartNotFoundIsSignalledWhenThereIsNoRestart() {
+    List<String> trail = new ArrayList<>();
+
+    try (Scope a = Scopes.create()) {
+      a.handle(BasicCondition.class, (s, ops) -> {
+        trail.add("BasicCondition");
+
+        return ops.restart(new UseValue<>(1));
+      });
+      a.handle(RestartNotFound.class, (s, ops) -> {
+        trail.add("RestartNotFound");
+
+        return ops.abort();
+      });
+
+      a.signal(new BasicCondition(""), new Policies<>(ReturnTypePolicy.ignore()));
+    } catch(AbortException a) {
+      assertLinesMatch(Arrays.asList("BasicCondition", "RestartNotFound"), trail);
+
+      return;
+    }
+
+    fail("Should've thrown an AbortException earlier!");
   }
 
   private String body(A a) {