+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Now there is a non-null root scope, providing a place to hang default handlers and restarts.
 * Several additions:
   * `HandlerNotFound`: A condition which `Scope.signal` uses instead of a policy. The root scope handles it by throwing a `HandlerNotFoundException`; but any scope may handle it as they choose. `Scope.notify`, for example, handles it by resuming execution. 
+  * `RestartNotFound`: A condition which `Handler.Operations.restart` uses instead of throwing directly. The root scope handles it by throwing a `RestartNotFoundException`; but any scope may handle it as they choose. 
 * Several removals:
   * `ScopeImpl.set`: `Scopes.create` now takes restarts, so it's no longer needed.
   * `ScopeImpl`: `Scope` is now a class, and therefore took over its duties.